// Client/eshop/src/api/models/AuthDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthDto
 */
export interface AuthDto {
    /**
     * 
     * @type {boolean}
     * @memberof AuthDto
     */
    isSucessfull?: boolean;
    /**
     * 
     * @type {string}
     * @memberof AuthDto
     */
    token?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthDto
     */
    errorMessage?: string;
}

export function AuthDtoFromJSON(json: any): AuthDto {
    return AuthDtoFromJSONTyped(json, false);
}

export function AuthDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'isSucessfull': !exists(json, 'isSucessfull') ? undefined : json['isSucessfull'],
        'token': !exists(json, 'token') ? undefined : json['token'],
        'errorMessage': !exists(json, 'errorMessage') ? undefined : json['errorMessage'],
    };
}

export function AuthDtoToJSON(value?: AuthDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'isSucessfull': value.isSucessfull,
        'token': value.token,
        'errorMessage': value.errorMessage,
    };
}